+++ internal/model/TypeStandardizationModel.go
@@ -1,6 +1,6 @@
 package model
 
 type TypeStandardizationModel struct {
-	Key string
-	Value int64
+	Key   string
+	Value int16
 }
