+++ internal/service/abstract/ICacheService.go
@@ -1,5 +1,5 @@
 package abstract
 
 type ICacheService interface {
-	ManageCache (tableName string, key string) (v int64, s bool, m string)
+	ManageCache(tableName string, key string) (v int16, s bool, m string)
 }
