+++ internal/service/concrete/CacheManager.go
@@ -11,37 +11,37 @@ import (
 )
 
 type CacheManager struct {
-	Cache *Cache.ICache
+	Cache                   *Cache.ICache
 	ITypeStandardizationDal *abstract.ITypeStandardizationDal
-	IJsonParser *Ijsonparser.IJsonParser
+	IJsonParser             *Ijsonparser.IJsonParser
 }
 
 func CacheManagerConstructor() *CacheManager {
 	return &CacheManager{Cache: &IoC.RedisCache,
 		ITypeStandardizationDal: &IoC.TypeStandardizationDal,
-		IJsonParser: &IoC.JsonParser,
-}
+		IJsonParser:             &IoC.JsonParser,
+	}
 }
 
-func (c *CacheManager) ManageCache(tableName string, key string) (v int64, s bool, m string){
+func (c *CacheManager) ManageCache(tableName string, key string) (v int16, s bool, m string) {
 
 	// cache sorgula,
 	value, err := (*c.Cache).Get(key)
-	if err != nil && err.Error() == "null data error"{
+	if err != nil && err.Error() == "null data error" {
 
 		//Bilgi yok ise veri tabanına sor
 		m, getErr := (*c.ITypeStandardizationDal).GetByKey(tableName, key)
-		if getErr != nil{
+		if getErr != nil {
 
 			log.Fatal("CacheManager",
 				"ManageCache",
 				"ITypeStandardizationDal_GetByKey", getErr.Error())
-			return  int64(0), false, getErr.Error()
+			return int16(0), false, getErr.Error()
 		}
 
 		//bu bilgi var ise veriyi dön ve cache'i güncelle,
-		if m!= nil{
-			_, err := (*c.Cache).Set(m.Key, m.Value, 10 )
+		if m != nil {
+			_, err := (*c.Cache).Set(m.Key, m.Value, 10)
 			if err != nil {
 				log.Fatal("CacheManager",
 					"ManageCache",
@@ -55,17 +55,17 @@ func (c *CacheManager) ManageCache(tableName string, key string) (v int64, s boo
 			log.Fatal("CacheManager",
 				"ManageCache",
 				"ITypeStandardizationDal_GetMaxByValue", maxErr.Error())
-			return int64(0), false, maxErr.Error()
+			return int16(0), false, maxErr.Error()
 		}
 
 		if err := (*c.ITypeStandardizationDal).Add(tableName, &model.TypeStandardizationModel{
 			Key:   key,
 			Value: max + 1,
-		}) ; err != nil {
+		}); err != nil {
 			log.Fatal("CacheManager",
 				"ManageCache",
 				"ITypeStandardizationDal_Add", err.Error())
-			return int64(0), false, err.Error()
+			return int16(0), false, err.Error()
 		}
 		return max + 1, true, ""
 	}
@@ -83,8 +83,7 @@ func (c *CacheManager) ManageCache(tableName string, key string) (v int64, s boo
 			"ManageCache",
 			"strconv", err.Error())
 
-		return int64(0), false, logErr.Error()
+		return int16(0), false, logErr.Error()
 	}
-	return int64(i), true, ""
+	return int16(i), true, ""
 }
-
